requirements-server.txt
# Core Dependencies
Django==4.2.7
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Werkzeug==3.0.1
WTForms==3.2.1

# Database
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.15.2

# AWS and Cloud Services
boto3==1.34.34
botocore==1.34.162
aiobotocore==2.21.1

# HTTP and API
aiohttp==3.11.16
requests==2.31.0
httpx==0.28.1

# Security
cryptography==44.0.2
PyJWT==2.10.1
argon2-cffi==23.1.0
bcrypt==4.3.0

# Data Processing
pandas==2.2.3
numpy==2.2.3
scipy==1.15.2

# Image Processing
Pillow==11.1.0
opencv-python-headless==4.11.0.86

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pdfplumber==0.11.6

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
tqdm==4.67.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.14.0

# Development Tools
black==23.7.0
flake8==6.1.0

# Custom Dependencies
-e git+https://github.com/chrscato/BRsystem@0410e8fa354178db3296640b4efaf8242b32a2ca#egg=brsystem 